Quit Node server on missing Github credentials
The Github API Key and User Agent string are crucial for the Github
invite system to work. Without them, any valid or invalid Github user
cannot be invited to the organization.

If both API Key and User Agent are not present, then display an error
message and exit the process.